working on  The requerments The user cannot log out. DONE There is no error message when the location permission is denied by the user. DONE The user's real-time location cannot be shown on the map when the foreground location permission is granted (Nexus 5X, API 29). DONE
Required permissions are checked when the user enters the save reminder screen. DONE

Reminder details are not displayed in the new screen. DONE

Live data objects are not tested using shouldReturnError testing function. DONE

Error handling is not done in the FakeDataSource class. DONE

Predictable errors like data not found are not tested. DONE